+++ resources/logical_enclosure.rb
@@ -11,6 +11,8 @@
 
 OneviewCookbook::ResourceBaseProperties.load(self)
 
+property :script, String # script for set_script action
+
 default_action :update_from_group
 
 action_class do
@@ -24,5 +26,36 @@ action :update_from_group do
   return if item['state'] == 'Consistent'
   converge_by "Update LogicalEnclosure '#{name}' from group" do
     item.update_from_group
+  end
+end
+
+action :reconfigure do
+  item = load_resource
+  item.retrieve!
+
+  item['enclosureUris'].each do |enclosure_uri|
+    enclosure = OneviewSDK::Enclosure.new(item.client, uri: enclosure_uri)
+    enclosure.retrieve!
+    next unless ['NotReapplyingConfiguration', 'ReapplyingConfigurationFailed', ''].include? enclosure['reconfigurationState']
+    converge_by "#{resource_name} '#{name}' was reconfigured." do
+      item.reconfigure
+    end
+    return true
+  end
+
+  Chef::Log.info("#{resource_name} '#{name}' is already reconfiguring.")
+end
+
+action :set_script do
+  item = load_resource
+  item.retrieve!
+
+  if item.get_script.eql? script
+    Chef::Log.info("#{resource_name} '#{name}' script is up to date")
+  else
+    Chef::Log.debug "#{resource_name} '#{name}' Chef resource differs from OneView resource."
+    converge_by "Updated script for #{resource_name} '#{name}'" do
+      item.set_script(script)
+    end
   end
 end